Added basic Error Handling for Homefragment
When there is an Error with the Request a Toast should be made
