+++ app/src/main/java/net/informatikag/thomapp/MainActivity.kt
@@ -1,5 +1,9 @@
 package net.informatikag.thomapp
 
+import android.app.Activity
+import android.content.Context
+import android.net.ConnectivityManager
+import android.net.NetworkInfo
 import android.os.Bundle
 import android.view.Menu
 import com.google.android.material.snackbar.Snackbar
@@ -12,7 +16,15 @@ import androidx.navigation.ui.setupWithNavController
 import androidx.drawerlayout.widget.DrawerLayout
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.android.volley.*
 import net.informatikag.thomapp.databinding.ActivityMainBinding
+import org.apache.http.conn.ConnectTimeoutException
+import org.json.JSONException
+import org.xmlpull.v1.XmlPullParserException
+import java.net.ConnectException
+import java.net.MalformedURLException
+import java.net.SocketException
+import java.net.SocketTimeoutException
 
 /**
  * This was mostly auto generated by Android Studio
@@ -23,6 +35,51 @@ class MainActivity : AppCompatActivity(){
     private lateinit var binding: ActivityMainBinding
 
     companion object {
+
+        /**
+         * Gets a displayable Error String from a VolleyError
+         * @param error the Error to generate from
+         * @param activity needed to get some Conenctivity Stats
+         */
+        fun getVolleyError(error: VolleyError?, activity: Activity): String {
+            var errorMsg = ""
+            if (error == null){
+                errorMsg = activity.getString(R.string.network_error_weired_response)
+            } else if (error is NoConnectionError) {
+                val cm = activity.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
+                var activeNetwork: NetworkInfo? = null
+                activeNetwork = cm.activeNetworkInfo
+                errorMsg = if (activeNetwork != null && activeNetwork.isConnectedOrConnecting) {
+                    activity.getString(R.string.network_error_server_error)
+                } else {
+                    activity.getString(R.string.network_error_not_connected)
+                }
+            } else if (error is NetworkError || error.cause is ConnectException) {
+                errorMsg = activity.getString(R.string.network_error_not_connected)
+            } else if (error.cause is MalformedURLException) {
+                errorMsg = activity.getString(R.string.network_error_weired_response)
+            } else if (error is ParseError || error.cause is IllegalStateException || error.cause is JSONException || error.cause is XmlPullParserException) {
+                errorMsg = activity.getString(R.string.network_error_weired_response)
+            } else if (error.cause is OutOfMemoryError) {
+                errorMsg = activity.getString(R.string.network_error_out_of_memory)
+            } else if (error is AuthFailureError) {
+                errorMsg = activity.getString(R.string.network_error_generic)
+            } else if (error is ServerError || error.cause is ServerError) {
+                activity.getString(R.string.network_error_server_error)
+            } else if (
+                error is TimeoutError ||
+                error.cause is SocketTimeoutException ||
+                error.cause is ConnectTimeoutException ||
+                error.cause is SocketException ||
+                (error.cause!!.message != null && error.cause!!.message!!.contains("Your connection has timed out, please try again"))
+            ) {
+                errorMsg = activity.getString(R.string.network_error_timeout)
+            } else {
+                errorMsg = activity.getString(R.string.network_error_generic)
+            }
+            return errorMsg
+        }
+
         // Substitution plan
         val SUBSTITUTION_PLAN_PREVIEW_AMOUNT = 10
         val SUBSTITUTION_PLAN_BASE_URL = "http://isiko404.codes:4000"   //TODO switch to HTTPS
