+++ app/src/main/java/net/informatikag/thomapp/utils/models/view/VertretungsplanViewModel.kt
@@ -1,11 +1,14 @@
 package net.informatikag.thomapp.utils.models.view
 
+import android.app.Activity
 import android.app.Application
 import android.content.Context
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
+import com.google.android.material.snackbar.Snackbar
 import net.informatikag.thomapp.MainActivity
+import net.informatikag.thomapp.R
 import net.informatikag.thomapp.utils.models.data.SubstitutionEntryData
 import net.informatikag.thomapp.utils.models.data.SubstitutionProfileData
 
@@ -20,7 +23,7 @@ class VertretungsplanViewModel(application: Application): AndroidViewModel(appli
 
     fun isEmpty():Boolean = _entrys.value == null
 
-    fun loadVertretunsplan(context: Context){
+    fun loadVertretunsplan(context: Context, mainActivity: Activity){
         // TODO replace this with letting the user Choose a Profile
         _profile.value = SubstitutionProfileData(
             -1,
@@ -36,8 +39,8 @@ class VertretungsplanViewModel(application: Application): AndroidViewModel(appli
         )
 
         _profile.value!!.getSubstitutionPlan(context){results, error ->
-            if (error != null){
-                // TODO add Error handling
+            if (error != null || results == null){
+                Snackbar.make(mainActivity.findViewById(R.id.app_bar_main), MainActivity.getVolleyError(error, mainActivity), Snackbar.LENGTH_LONG).show()
             } else {
                 for (entry in results!!.entryData) {
                     if (!isEmpty())
